Seperated cogs into their own files
All of the individual cogs have been moved out of cogs.py, and into their own directory. cogs.py simply imports the class definitions from these files, uses those in a function to add them to the bot. scripts.py was also removed during this process, with it's functions being absorbed into their corresponding cog files. Some functions have received minor tweaks as this integration happened.
